Place and style `ThemeControl`
