+++ src/blueprintjsIcons.js
@@ -1,6 +1,13 @@
 import { IconNames, IconSvgPaths16 as BI16, IconSvgPaths20 as BI20 } from '@blueprintjs/icons'
 
-const usedIcons = [IconNames.TICK_CIRCLE, IconNames.CARET_DOWN, IconNames.UNLOCK, IconNames.LOCK]
+const usedIcons = [
+	IconNames.TICK_CIRCLE,
+	IconNames.CARET_DOWN,
+	IconNames.UNLOCK,
+	IconNames.LOCK,
+	IconNames.FLASH,
+	IconNames.MOON,
+]
 
 const IconSvgPaths16 = {}
 const IconSvgPaths20 = {}
