+++ src/components/NavBar/ThemeControl.jsx
@@ -1,0 +1,30 @@
+import { Button, ButtonGroup, Colors } from '@blueprintjs/core'
+import { IconNames } from '@blueprintjs/icons'
+import { css } from 'linaria'
+import React, { useState } from 'react'
+
+export const ThemeControl = () => {
+	const [selectedTheme, changeTheme] = useState('light')
+	const isLightTheme = selectedTheme === 'light'
+
+	return (
+		<ButtonGroup
+			className={css`
+				margin-left: 23%;
+			`}
+		>
+			<Button
+				active={isLightTheme}
+				intent={isLightTheme ? 'primary' : 'none'}
+				icon={IconNames.FLASH}
+				onClick={() => changeTheme('light')}
+			/>
+			<Button
+				active={!isLightTheme}
+				intent={isLightTheme ? 'none' : 'primary'}
+				icon={IconNames.MOON}
+				onClick={() => changeTheme('dark')}
+			/>
+		</ButtonGroup>
+	)
+}
