--- src/components/NavBar/index.js
@@ -1,1 +1,0 @@
-export { NavigationBar } from './NavBar'
